Create dummy test suite
